fixed setup

+++ actions/user-actions.php
@@ -1,6 +1,6 @@
 <?php
-require_once '../models/Teacher.php';
-require_once '../services/TeacherService.php';
+require_once '../models/User.php';
+require_once '../services/UserService.php';
 require_once '../common/Response.php';
 require_once '../services/UploadService.php';
 require_once '../config/Config.php';
@@ -9,19 +9,35 @@ global $_CONFIG;
 $uploadService = new UploadService($_CONFIG["UPLOADDIR"]);
 $response = new Response();
 
-#deletes a teacher by the ID posted
+#deletes a user by the ID posted
 if(isset($_GET['id']) && isset($_GET['action'])){
-    $teacher = new Teacher();
-    $teacher->setId($_GET['id']);
-    $result = TeacherService::delete($teacher);
+    $user = new User();
+    $user->setId($_GET['id']);
+    $users = UserService::findAll();
+    $count = 0;
+    foreach($users as $user){
+        if($user->getRole() == "Admin"){
+            $count++;
+        }
+    }
+    
+    if($count == 1){
+        $response->code = 41;
+        $response->status = "Denied";
+        $response->message = "Cannot delete";
+        header('Location: ../templates/user-view.php?code='.$response->code);
+        exit;
+    }
+    $result = UserService::delete($user);
+   
     if($result["status"]){
         if(strcmp($result["path"], "") != 0) { $uploadService->removeFile($result["path"]); }
         $response->code = 200;
         $response->status = "Ok";
-        $response->message = "Teacher deleted successfully";
+        $response->message = "user deleted successfully";
         echo json_encode($response);
         if(!isset($_GET['ajax'])){
-          header('Location: ../templates/teacher-view.php?code='.$response->code);
+          header('Location: ../templates/user-view.php?code='.$response->code);
         }
         exit;
     };
@@ -35,21 +51,20 @@ if(strcmp($_FILES['file']['name'], "") != 0){
 #get upload results
 $file = (isset($result))? $result: ["status"=>"ok", "uploadedFile"=>""];
 
-#creates a teacher object from the $_POST & $_FILE arrays
-$teacher = TeacherService::getFromPost($_POST, $file["uploadedFile"]);
+#create a user object from the $_POST & $_FILE arrays
+$user = UserService::getUserFromPost($_POST, $file["uploadedFile"]);
 
-
-#updates a teacher's info
+#updates a user's info
 if(isset($_POST['id'])){
     if($_POST['id'] !=""){
-        $teacher->setId($_POST['id']);
-        if(TeacherService::update($teacher)){
+        $user->setId($_POST['id']);
+        if(UserService::update($user)){
             $response->code = 200;
             $response->status = "Ok";
-            $response->message = "Teacher updated successfully";
+            $response->message = "user updated successfully";
             echo json_encode($response);
             if(!isset($_GET['ajax'])){
-              header('Location: ../templates/teacher-form.php?code='.$response->code);
+              header('Location: ../templates/user-form.php?code='.$response->code);
             }
             exit;
         }else{
@@ -58,32 +73,32 @@ if(isset($_POST['id'])){
             $response->message = "a server error has occured";
             $uploadService->removeFiles();
             echo json_encode($response);
-            header('Location: ../templates/teacher-form.php?code='.$response->code);
+            header('Location: ../templates/user-form.php?code='.$response->code);
             exit;
         }  
     }
 }
 
-#check if teacher already exist
-if(TeacherService::exist($teacher)){
+#check if user already exist
+if(UserService::exist($user)){
     $response->code = 409;
     $response->status = "Failed";
     $response->message = "Duplicate entry";
     echo json_encode($response);
     if(!isset($_GET['ajax'])){
-      header('Location: ../templates/teacher-form.php?code='.$response->code);
+      header('Location: ../templates/user-form.php?code='.$response->code);
     }
     exit;
 }
 
-#creates a new teacher's record in the database
-if(TeacherService::insert($teacher)){
+#creates a new user's record in the database
+if(UserService::insert($user)){
     $response->code = 200;
     $response->status = "Ok";
-    $response->message = "Teacher inserted successfully";
+    $response->message = "user inserted successfully";
     echo json_encode($response);
     if(!isset($_GET['ajax'])){
-        header('Location: ../templates/teacher-form.php?code='.$response->code);
+      header('Location: ../templates/user-form.php?code='.$response->code);
     }
     exit;
  }else{
@@ -92,5 +107,5 @@ if(TeacherService::insert($teacher)){
     $response->message = "a server error has occured";
     $uploadService->removeFiles();
     echo json_encode($response);
-    header('Location: ../templates/teacher-form.php?code='.$response->code);
+    header('Location: ../templates/user-form.php?code='.$response->code);
  }
